+++ NEWS.md
@@ -1,6 +1,7 @@
 ### 0.18.0 (...)
 
 * Added Layer via Copy/Cut actions
+* Added support for Paste in Place action for tile layers
 * Added context menu to change custom property type (by Dmitry Hrabrov)
 * Added support for higher precision for custom floating point properties
 * Added %mappath variable to commands (by Jack Roper)